Add function to help with bi-directional encoding; code cleanup.
The encoder can do bi-directional, but the decoder cannot.
So we have a bidirectional encoder with a one-directional decoder; but then the
sizes of the hidden states are different.
To fix this, use double the hidden state size in the decoder, but we should
probaly reorder the dimensions a bit. That happens in reshape_hidden